cleaned up dependencys MVC
