--- src/main/scala/de/htwg/se/Chess/Controller/SetCommand.scala
@@ -1,16 +1,0 @@
-package de.htwg.se.Chess.controller
-
-import de.htwg.se.Chess.util.Command
-
-class SetCommand(pos_old:String, pos_new:String, controller: Controller) extends Command {
-
-  override def doMove: Unit =
-    controller.field = controller.field.move(pos_old, pos_new)
-
-  override def undoMove: Unit =
-    controller.field = controller.field.move(pos_old, pos_new)
-
-  override def redoMove: Unit =
-    controller.field = controller.field.move(pos_old, pos_new)
-
-}
