--- src/main/scala/de/htwg/se/Chess/util/Command.scala
@@ -1,7 +1,0 @@
-package de.htwg.se.Chess.util
-
-trait Command {
-  def doMove:Unit
-  def undoMove:Unit
-  def redoMove:Unit
-}
