--- src/main/scala/de/htwg/se/Chess/util/UndoManager.scala
@@ -1,31 +1,0 @@
-package de.htwg.se.Chess.util
-
-class UndoManager {
-  private var undoStack: List[Command]= Nil
-  private var redoStack: List[Command]= Nil
-  def doMove(command: Command) = {
-    undoStack = command::undoStack
-    command.doMove
-  }
-
-  def undoMove = {
-    undoStack match {
-      case  Nil =>
-      case head::stack => {
-        head.undoMove
-        undoStack=stack
-        redoStack= head::redoStack
-      }
-    }
-  }
-  def redoMove = {
-    redoStack match {
-      case Nil =>
-      case head::stack => {
-        head.redoMove
-        redoStack=stack
-        undoStack=head::undoStack
-      }
-    }
-  }
-}
